+++ Project/Code/src/Engine/Time.h
@@ -3,7 +3,7 @@ namespace Engine{
 	class Time
 	{
 	public:
-		static unsigned long getTime();
+		static double getTime();
 	};
 }
